refactor(devtools): migrate cache panel to cache and panel files
